requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-multiforloop==0.3.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2-binary==2.9.6
python-slugify==8.0.1
PyYAML==6.0
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
typing_extensions==4.6.3
xlrd==2.0.1
xlwt==1.3.0
